Connect the footer stuff further down.
